Update to work with GameLib after Asteroids update

+++ Pong/src/edu/elon/honors/price/pong2/PongData.java
@@ -7,9 +7,9 @@ public class PongData implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 	private int player1Score, player2Score;
-	private double paddle1X, paddle1Y, paddle2X, paddle2Y;
+	private float paddle1X, paddle1Y, paddle2X, paddle2Y;
 
-	private double ballVelX, ballVelY, ballX, ballY;
+	private float ballVelX, ballVelY, ballX, ballY;
 	
 	public int getPlayer1Score() {
 		return player1Score;
@@ -27,67 +27,67 @@ public class PongData implements Serializable {
 		this.player2Score = player2Score;
 	}
 
-	public double getPaddle1X() {
+	public float getPaddle1X() {
 		return paddle1X;
 	}
 
-	public void setPaddle1X(double paddle1x) {
+	public void setPaddle1X(float paddle1x) {
 		paddle1X = paddle1x;
 	}
 
-	public double getPaddle1Y() {
+	public float getPaddle1Y() {
 		return paddle1Y;
 	}
 
-	public void setPaddle1Y(double paddle1y) {
+	public void setPaddle1Y(float paddle1y) {
 		paddle1Y = paddle1y;
 	}
 
-	public double getPaddle2X() {
+	public float getPaddle2X() {
 		return paddle2X;
 	}
 
-	public void setPaddle2X(double paddle2x) {
+	public void setPaddle2X(float paddle2x) {
 		paddle2X = paddle2x;
 	}
 
-	public double getPaddle2Y() {
+	public float getPaddle2Y() {
 		return paddle2Y;
 	}
 
-	public void setPaddle2Y(double paddle2y) {
+	public void setPaddle2Y(float paddle2y) {
 		paddle2Y = paddle2y;
 	}
 
-	public double getBallVelX() {
+	public float getBallVelX() {
 		return ballVelX;
 	}
 
-	public void setBallVelX(double ballVelX) {
+	public void setBallVelX(float ballVelX) {
 		this.ballVelX = ballVelX;
 	}
 
-	public double getBallVelY() {
+	public float getBallVelY() {
 		return ballVelY;
 	}
 
-	public void setBallVelY(double ballVelY) {
+	public void setBallVelY(float ballVelY) {
 		this.ballVelY = ballVelY;
 	}
 
-	public double getBallX() {
+	public float getBallX() {
 		return ballX;
 	}
 
-	public void setBallX(double ballX) {
+	public void setBallX(float ballX) {
 		this.ballX = ballX;
 	}
 
-	public double getBallY() {
+	public float getBallY() {
 		return ballY;
 	}
 
-	public void setBallY(double ballY) {
+	public void setBallY(float ballY) {
 		this.ballY = ballY;
 	}
 }
